requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-CORS==4.0.0
Werkzeug==2.3.7
python-dotenv==1.0.0
psycopg2-binary==2.9.7
Flask-Mail==0.9.1
Pillow==10.0.1
